reorganized file structure for routes, added in health metrics, changed date input from number to string
