--- database/UserProfile/index.js
@@ -1,14 +1,0 @@
-const mongoose = require('mongoose');
-
-const mongoURI = 'mongodb://localhost:27017/bestEatsInc';
-
-const db = mongoose.connect(mongoURI, { useNewUrlParser: true });
-
-db
-  .then(db => console.log(`Connected to: ${mongoURI}`))
-  .catch(err => {
-    console.log(`There was a problem connecting to mongo at: ${mongoURI}`);
-    console.log(err);
-  });
-
-module.exports = db;
