+++ server/User/controller/createUser.js
@@ -1,4 +1,4 @@
-const User = require('../../../database/UserProfile/model/users');
+const User = require('../../../database/model/users');
 
 module.exports = {
   post: (req, res) => {
